+++ packages/core/config/services.php
@@ -21,7 +21,8 @@ return static function (ContainerConfigurator $container): void {
             ->exclude([
                 __DIR__ . '/../src/Entity/',
                 __DIR__ . '/../src/Module/{ModuleBlueprint,CollectionBlueprint,FieldBlueprint}.php',
-                __DIR__ . '/../src/Record/Record.php',
+                __DIR__ . '/../src/Money/Amount.php',
+                __DIR__ . '/../src/Record/{Record,Derivation}.php',
                 __DIR__ . '/../src/Validation/UniqueFieldValue.php',
                 __DIR__ . '/../src/XiviCoreBundle.php',
             ]);